requirements.txt
streamlit
pandas
google-cloud-bigquery
google-auth
pyarrow
db-dtypes